+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added required sign for hours title in Event2SetupSurveyPanel [#3582](https://github.com/rokwire/illinois-app/issues/3582).
 - Added default course detail image [#2286](https://github.com/rokwire/illinois-app/issues/2286).
 - Handle sport events from Calendar BB [#3594](https://github.com/rokwire/illinois-app/issues/3594).
+### Changed
+- More precise processing of 'Confirm URL' command in Event2CreatePanel [#3580](https://github.com/rokwire/illinois-app/issues/3580).
 
 ## [5.0.50] - 2023-08-18
 ### Changed